feat: Add manual adjustment for total number of issuers
Introduces a new field `totalEmisoresAjustes` to `InvoiceConfig` to allow administrators to manually set the total number of issuers for a given month and user. This provides more flexibility in reporting and billing calculations beyond the automatically derived count.

Also adds the `Trash2` icon to the available icons in `Factura.tsx`.